pyproject.toml
[tool.poetry]
name = "kruiscijferraadsel"
version = "0.1.0"
description = ""
authors = ["Joris Boeye <joris.boeye@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
attrs = "^21.2.0"
numpy = "^1.21.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^21.10b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
